+++ app/models/e_mail.rb
@@ -1,5 +1,6 @@
 class EMail < ApplicationRecord
   include Auditable
+  include Emails::Fetch
   include Emails::Search
   include Emails::Scopes
   include Emails::Validations
